feature+technical<all layers>: create book service and controller with basic handling of exceptions + add mapstruct

+++ repository/src/main/java/org/thibaut/thelibrary/repository/repository/BookRepository.java
@@ -9,7 +9,7 @@ import org.thibaut.thelibrary.entity.BookEntity;
 import java.util.List;
 
 @Repository
-public interface BookRepository extends JpaRepository< BookEntity, Integer > {
+public interface BookRepository extends JpaRepository< BookEntity, Long > {
 
 
 	List< BookEntity > getBooksByTitleContains( String title);
